requirements.txt
instructor~=1.5.2
openai~=1.51.2
python-dotenv~=1.0.1
requests~=2.32.3
flask~=3.0.3
flask-cors~=4.0.1
pydantic~=2.9.2
attrs~=23.2.0
tomli~=2.0.1
redis~=5.1.0
fakeredis~=2.25.1
pytest~=8.3.3
keycloak=~3.1.5
tenacity~=8.5.0
unstructured~=0.16.0
pandas~=2.2.3
pydantic_core~=2.23.4
keycloak=~3.1.5
tenacity~=8.5.0
unstructured~=0.16.0
pandas~=2.2.3
pydantic_core~=2.23.4